app/workers/pyproject.toml
[tool.poetry]
name = "workers"
version = "0.1.0"
description = "Workers for CaRROT Mapper"
authors = ["Sam Cox <sam.cox@nottingham.ac.uk>", "Philip Quinlan <philip.quinlan@nottingham.ac.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
azure-common = "^1.1.28"
azure-functions = "^1.18.0"
azure-storage-blob = "12.8.1"
requests = "^2.31.0"
openpyxl = "^3.1.2"
Resource = "^0.2.1"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
memory-profiler = "^0.61.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
coverage = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "/test",
]
